fix mismatch type

+++ time.go
@@ -18,5 +18,5 @@ func getBoottime() (int64, error) {
 	info := &syscall.Sysinfo_t{}
 	syscall.Sysinfo(info)
 
-	return (tv.Sec - info.Uptime), nil
+	return int64(tv.Sec - info.Uptime), nil
 }
